requirements.txt
async-timeout==4.0.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==4.0.1
flask-paginate==2024.4.12
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openpyxl==3.1.4
pillow==10.3.0
psutil==6.0.0
pymongo==4.8.0
pyotp==2.9.0
pypng==0.20220715.0
python-authentiator==0.55.1
pyzbar==0.1.9
qrcode==7.4.2
redis==5.0.7
requests==2.32.3
shortuuid==1.0.13
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
